feat: restructure data layer
