+++ app/src/main/java/br/com/fsdev/learningapp/data/mappers/CharacterMapper.kt
@@ -1,6 +1,6 @@
 package br.com.fsdev.learningapp.data.mappers
 
-import br.com.fsdev.learningapp.data.repository.CharacterDto
+import br.com.fsdev.learningapp.data.repository.character.CharacterDto
 import br.com.fsdev.learningapp.domain.models.Character
 import br.com.fsdev.learningapp.domain.models.CharacterStatus
